+++ rust/candid/src/parser/value.rs
@@ -1,16 +1,13 @@
 use super::token::error;
 use super::typing::TypeEnv;
-use crate::types::number::{pp_num_str, Int, Nat};
+use crate::types::number::{Int, Nat};
 use crate::types::{Field, Label, Type};
 use crate::{Error, Result};
-use arbitrary::Unstructured;
 use serde::de;
 use serde::de::{Deserialize, Visitor};
 use std::collections::HashMap;
 use std::fmt;
 use std::ops::Deref;
-
-const MAX_ELEMENTS_FOR_PRETTY_PRINT: usize = 10;
 
 #[derive(PartialEq, Clone)]
 pub enum IDLValue {
@@ -66,14 +63,6 @@ impl IDLArgs {
         let mut args = Vec::new();
         for (v, ty) in self.args.iter().zip(types.iter()) {
             let v = v.annotate_type(from_parser, env, &ty)?;
-            args.push(v);
-        }
-        Ok(IDLArgs { args })
-    }
-    pub fn any(u: &mut Unstructured, env: &TypeEnv, types: &[Type]) -> Result<Self> {
-        let mut args = Vec::new();
-        for t in types.iter() {
-            let v = IDLValue::any(u, env, t)?;
             args.push(v);
         }
         Ok(IDLArgs { args })
@@ -139,22 +128,6 @@ impl std::str::FromStr for IDLValue {
     fn from_str(str: &str) -> std::result::Result<Self, Self::Err> {
         let lexer = super::token::Tokenizer::new(str);
         Ok(super::grammar::ArgParser::new().parse(lexer)?)
-    }
-}
-
-impl fmt::Display for IDLArgs {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", pretty::pp_args(&self).pretty(80))
-    }
-}
-
-impl fmt::Display for IDLValue {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "{}",
-            pretty::pp_value(MAX_ELEMENTS_FOR_PRETTY_PRINT, &self).pretty(80)
-        )
     }
 }
 
@@ -340,254 +313,6 @@ impl IDLValue {
                 Type::Func(f)
             }
         }
-    }
-    pub fn any(u: &mut Unstructured, env: &TypeEnv, ty: &Type) -> Result<Self> {
-        Ok(match ty {
-            Type::Var(id) => {
-                let ty = env.rec_find_type(id)?;
-                Self::any(u, env, &ty)?
-            }
-            Type::Knot(ref id) => {
-                let ty = crate::types::internal::find_type(id).unwrap();
-                Self::any(u, env, &ty)?
-            }
-            Type::Null => IDLValue::Null,
-            Type::Bool => IDLValue::Bool(u.arbitrary()?),
-            Type::Nat => IDLValue::Nat(u.arbitrary::<u128>()?.into()),
-            Type::Int => IDLValue::Int(u.arbitrary::<i128>()?.into()),
-            Type::Nat8 => IDLValue::Nat8(u.arbitrary()?),
-            Type::Nat16 => IDLValue::Nat16(u.arbitrary()?),
-            Type::Nat32 => IDLValue::Nat32(u.arbitrary()?),
-            Type::Nat64 => IDLValue::Nat64(u.arbitrary()?),
-            Type::Int8 => IDLValue::Int8(u.arbitrary()?),
-            Type::Int16 => IDLValue::Int16(u.arbitrary()?),
-            Type::Int32 => IDLValue::Int32(u.arbitrary()?),
-            Type::Int64 => IDLValue::Int64(u.arbitrary()?),
-            Type::Float32 => IDLValue::Float32(u.arbitrary()?),
-            Type::Float64 => IDLValue::Float64(u.arbitrary()?),
-            Type::Text => IDLValue::Text(u.arbitrary()?),
-            Type::Reserved => IDLValue::Reserved,
-            Type::Principal => IDLValue::Principal(crate::Principal::anonymous()),
-            Type::Opt(t) => {
-                if u.arbitrary::<bool>()? {
-                    IDLValue::Opt(Box::new(Self::any(u, env, t)?))
-                } else {
-                    IDLValue::None
-                }
-            }
-            Type::Vec(t) => {
-                // TODO: use arbitrary_len or a config
-                let len = u.int_in_range(0..=50)?;
-                let mut vec = Vec::with_capacity(len);
-                for _ in 0..len {
-                    let e = Self::any(u, env, t)?;
-                    vec.push(e);
-                }
-                IDLValue::Vec(vec)
-            }
-            Type::Record(fs) => {
-                let mut res = Vec::new();
-                for Field { id, ty } in fs.iter() {
-                    let val = Self::any(u, env, ty)?;
-                    res.push(IDLField {
-                        id: id.clone(),
-                        val,
-                    });
-                }
-                IDLValue::Record(res)
-            }
-            Type::Variant(fs) => {
-                let idx = u.int_in_range(0..=(fs.len() - 1))?;
-                let Field { id, ty } = &fs[idx];
-                let val = Self::any(u, env, ty)?;
-                let field = IDLField {
-                    id: id.clone(),
-                    val,
-                };
-                IDLValue::Variant(Box::new(field), idx as u64)
-            }
-            _ => unimplemented!(),
-        })
-    }
-}
-
-pub mod pretty {
-    use super::*;
-    use crate::pretty::*;
-
-    use ::pretty::RcDoc;
-
-    pub use crate::bindings::candid::pp_label;
-
-    // The definition of tuple is language specific.
-    pub(crate) fn is_tuple(t: &IDLValue) -> bool {
-        match t {
-            IDLValue::Record(ref fs) => {
-                for (i, field) in fs.iter().enumerate() {
-                    if field.id.get_id() != (i as u32) {
-                        return false;
-                    }
-                }
-                true
-            }
-            _ => false,
-        }
-    }
-
-    fn pp_field(depth: usize, field: &IDLField, is_variant: bool) -> RcDoc {
-        let val_doc = if is_variant && field.val == IDLValue::Null {
-            RcDoc::nil()
-        } else {
-            kwd(" =").append(pp_value(depth - 1, &field.val))
-        };
-        pp_label(&field.id).append(val_doc)
-    }
-
-    fn pp_fields(depth: usize, fields: &[IDLField]) -> RcDoc {
-        let fs = concat(fields.iter().map(|f| pp_field(depth, f, false)), ";");
-        enclose_space("{", fs, "}")
-    }
-
-    pub fn pp_char(v: u8) -> String {
-        if (0x20..=0x7e).contains(&v) && v != 0x22 && v != 0x5c {
-            std::char::from_u32(v as u32).unwrap().to_string()
-        } else {
-            format!("\\{:02x}", v)
-        }
-    }
-
-    pub fn pp_value(depth: usize, v: &IDLValue) -> RcDoc {
-        use super::IDLValue::*;
-        if depth == 0 {
-            return RcDoc::as_string(format!("{:?}", v));
-        }
-        match v {
-            Text(ref s) => RcDoc::as_string(format!("\"{}\"", s)),
-            Opt(v) => kwd("opt").append(pp_value(depth - 1, v)),
-            Vec(vs) => {
-                if let Some(Nat8(_)) = vs.first() {
-                    RcDoc::as_string(format!("{:?}", v))
-                } else if vs.len() > MAX_ELEMENTS_FOR_PRETTY_PRINT {
-                    RcDoc::as_string(format!("{:?}", v))
-                } else {
-                    let body = concat(vs.iter().map(|v| pp_value(depth - 1, v)), ";");
-                    kwd("vec").append(enclose_space("{", body, "}"))
-                }
-            }
-            Record(fields) => {
-                if is_tuple(v) {
-                    let tuple = concat(fields.iter().map(|f| pp_value(depth - 1, &f.val)), ";");
-                    kwd("record").append(enclose_space("{", tuple, "}"))
-                } else {
-                    kwd("record").append(pp_fields(depth, &fields))
-                }
-            }
-            Variant(v, _) => {
-                kwd("variant").append(enclose_space("{", pp_field(depth, &v, true), "}"))
-            }
-            _ => RcDoc::as_string(format!("{:?}", v)),
-        }
-    }
-
-    pub fn pp_args(args: &IDLArgs) -> RcDoc {
-        let body = concat(
-            args.args
-                .iter()
-                .map(|v| pp_value(MAX_ELEMENTS_FOR_PRETTY_PRINT, v)),
-            ",",
-        );
-        enclose("(", body, ")")
-    }
-}
-
-impl fmt::Debug for IDLArgs {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        if self.args.len() == 1 {
-            write!(f, "({:?})", self.args[0])
-        } else {
-            let mut tup = f.debug_tuple("");
-            for arg in self.args.iter() {
-                tup.field(arg);
-            }
-            tup.finish()
-        }
-    }
-}
-impl fmt::Debug for IDLValue {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use IDLValue::*;
-        match self {
-            Null => write!(f, "null"),
-            Bool(b) => write!(f, "{}", b),
-            Number(n) => write!(f, "{}", n),
-            Int(i) => write!(f, "{}", i),
-            Nat(n) => write!(f, "{}", n),
-            Nat8(n) => write!(f, "{}", n),
-            Nat16(n) => write!(f, "{}", pp_num_str(&n.to_string())),
-            Nat32(n) => write!(f, "{}", pp_num_str(&n.to_string())),
-            Nat64(n) => write!(f, "{}", pp_num_str(&n.to_string())),
-            Int8(n) => write!(f, "{}", n),
-            Int16(n) => write!(f, "{}", pp_num_str(&n.to_string())),
-            Int32(n) => write!(f, "{}", pp_num_str(&n.to_string())),
-            Int64(n) => write!(f, "{}", pp_num_str(&n.to_string())),
-            Float32(n) => write!(f, "{}", n),
-            Float64(n) => write!(f, "{}", n),
-            Text(s) => write!(f, "{:?}", s),
-            None => write!(f, "null"),
-            Reserved => write!(f, "reserved"),
-            Principal(id) => write!(f, "principal \"{}\"", id),
-            Service(id) => write!(f, "service \"{}\"", id),
-            Func(id, meth) => write!(
-                f,
-                "func \"{}\".{}",
-                id,
-                crate::bindings::candid::ident_string(meth)
-            ),
-            Opt(v) => write!(f, "opt {:?}", v),
-            Vec(vs) => {
-                if let Some(Nat8(_)) = vs.first() {
-                    write!(f, "blob \"")?;
-                    for v in vs.iter() {
-                        match v {
-                            Nat8(v) => write!(f, "{}", &pretty::pp_char(*v))?,
-                            _ => unreachable!(),
-                        }
-                    }
-                    write!(f, "\"")
-                } else {
-                    write!(f, "vec {{")?;
-                    for v in vs.iter() {
-                        write!(f, " {:?};", v)?
-                    }
-                    write!(f, "}}")
-                }
-            }
-            Record(fs) => {
-                write!(f, "record {{")?;
-                for (i, e) in fs.iter().enumerate() {
-                    if e.id.get_id() == i as u32 {
-                        write!(f, " {:?};", e.val)?;
-                    } else {
-                        write!(f, " {:?};", e)?;
-                    }
-                }
-                write!(f, "}}")
-            }
-            Variant(v, _) => {
-                write!(f, "variant {{ ")?;
-                if v.val == Null {
-                    write!(f, "{}", v.id)?;
-                } else {
-                    write!(f, "{:?}", v)?;
-                }
-                write!(f, " }}")
-            }
-        }
-    }
-}
-impl fmt::Debug for IDLField {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{} = {:?}", self.id, self.val)
     }
 }
